webdav: fix a missing typecast
Patch: http://rb.dcache.org/r/2316/

+++ modules/dCache/org/dcache/webdav/DcacheResourceFactory.java
@@ -628,7 +628,7 @@ public class DcacheResourceFactory
                                    e.toString());
             throw e;
         } finally {
-            _downloads.remove(transfer.getSessionId());
+            _downloads.remove((int) transfer.getSessionId());
             _transfers.remove(transfer);
         }
     }
